requirements.txt
# Core dependencies
pip>=23.3; sys_platform == 'darwin'
wheel; sys_platform == 'darwin'
PyYAML; sys_platform == 'darwin'
numpy==1.23.5
requests>=2.31.0,<2.32.0
tqdm
wget

# Audio processing
ffmpeg-python>=0.2.0
faiss-cpu==1.7.3
librosa==0.9.2
scipy==1.11.1
soundfile==0.12.1
noisereduce
pedalboard
stftpitchshift

# Machine learning and deep learning
omegaconf>=2.0.6; sys_platform == 'darwin'
numba; sys_platform == 'linux'
numba==0.57.0; sys_platform == 'darwin' or sys_platform == 'win32'
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
torchcrepe==0.0.23
torchfcpe
einops
libf0
transformers==4.44.2

# Visualization and UI
matplotlib==3.7.2
tensorboard
gradio==5.5.0

# Miscellaneous utilities
certifi>=2023.07.22; sys_platform == 'darwin'
antlr4-python3-runtime==4.8; sys_platform == 'darwin'
tensorboardX
edge-tts==6.1.9
pypresence
beautifulsoup4
flask
google-cloud-storage
